+++ Microsoft.PowerApps.UIAutomation.Sample/ModelDriven/RunSolutionChecker.cs
@@ -41,6 +41,8 @@ namespace Microsoft.PowerApps.UIAutomation.Sample.ModelDriven
 
         }
 
+        [TestCategory("SolutionCheckerAutomation")]
+        [Priority(1)]
         [TestMethod]
         public void TestRunSolutionCheckerFromCommandBar()
         {
@@ -168,6 +170,8 @@ namespace Microsoft.PowerApps.UIAutomation.Sample.ModelDriven
             }
         }
 
+        [TestCategory("SolutionCheckerAutomation")]
+        [Priority(1)]
         [TestMethod]
         public void TestRunSolutionCheckerFromSolutionsGrid()
         {
@@ -291,6 +295,74 @@ namespace Microsoft.PowerApps.UIAutomation.Sample.ModelDriven
             }
         }
 
+        [TestCategory("SolutionCheckerAutomation")]
+        [Priority(1)]
+        [TestMethod]
+        public void TestVerifySolutionChecker()
+        {
+            BrowserOptions options = TestSettings.Options;
+            options.BrowserType = _browserType;
+
+            using (var appBrowser = new PowerAppBrowser(TestSettings.Options))
+            {
+                try
+                {
+                    // Login To PowerApps
+                    Console.WriteLine("Performing Login");
+
+                    for (int retryCount = 0; retryCount < Reference.Login.SignInAttempts; retryCount++)
+                    {
+                        try
+                        {
+                            appBrowser.OnlineLogin.Login(_xrmUri, _username.ToSecureString(), _password.ToSecureString());
+                            break;
+                        }
+                        catch (Exception)
+                        {
+                            appBrowser.Navigate("about:blank");
+                            if (retryCount == Reference.Login.SignInAttempts)
+                            {
+                                Console.WriteLine("Login failed after {0} attempts.", retryCount);
+                                throw;
+                            }
+                            Console.WriteLine("Login failed in #{0} attempt.", retryCount);
+                            continue;
+                        }
+                    }
+                    Console.WriteLine("Login Complete");
+
+                    //Pick the Org
+                    Console.WriteLine($"Changing PowerApps Environment to {_environmentName}");
+                    appBrowser.Navigation.ChangeEnvironment(_environmentName);
+                    appBrowser.ThinkTime(1500);
+
+                    //Click Solutions
+                    Console.WriteLine($"Click {_sideBarButton} via Sidebar");
+                    appBrowser.SideBar.Navigate(_sideBarButton);
+
+                    Console.WriteLine("Make sure each managed solution does not have the solution checker button in the command bar");
+                    appBrowser.ModelDrivenApps.VerifyManagedSolutionsUnavailable(_commandBarButton, 1000);
+
+                    appBrowser.ThinkTime(5000);
+
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine($"An error occurred during Solution Checker test run for solution {_solutionName}: {e}");
+                    string location = $@"{_resultsDirectory}\RunSolutionChecker-{_solutionName}-GenericError.bmp";
+
+                    appBrowser.TakeWindowScreenShot(location, OpenQA.Selenium.ScreenshotImageFormat.Bmp);
+                    _testContext.AddResultFile(location);
+
+                    Assert.Fail($"An error occurred during Solution Checker test run for solution {_solutionName}: {e}");
+                }
+
+                Console.WriteLine("Solution Checker Test Run Complete");
+            }
+        }
+
+        [TestCategory("SolutionCheckerAutomation")]
+        [Priority(2)]
         [TestMethod]
         public void TestCancelSolutionCheckerRun()
         {
@@ -368,16 +440,19 @@ namespace Microsoft.PowerApps.UIAutomation.Sample.ModelDriven
                     // Get solution check status post-cancel
                     string solutionCheckStatus = appBrowser.ModelDrivenApps.GetCurrentStatus(_solutionName);
 
-                    // Validate that we did not receive an error status
-                    if (!solutionCheckStatus.Contains("Results as of"))
+                    if (appBrowser.Options.BrowserType.ToString() != "IE")
                     {
-                        throw new ApplicationException($"Unexpected Solution Check Status. Value '{solutionCheckStatus}' received instead of 'Results as of ...' ");
-                    }
+                        // Validate that we did not receive an error status
+                        if (!solutionCheckStatus.Contains("Results as of"))
+                        {
+                            throw new ApplicationException($"Unexpected Solution Check Status. Value '{solutionCheckStatus}' received instead of 'Results as of ...' ");
+                        }
 
-                    // Validate that the cancel was successful and the "new" status is equal to the original starting status
-                    if (solutionCheckStatus != originalSolutionStatus)
-                    {
-                        throw new ApplicationException($"Unexpected Solution Check Status. Value '{solutionCheckStatus}' has changed from '{originalSolutionStatus}'. A failure occurred during cancellation and was not reported to the message bar. ");
+                        // Validate that the cancel was successful and the "new" status is equal to the original starting status
+                        if (solutionCheckStatus != originalSolutionStatus)
+                        {
+                            throw new ApplicationException($"Unexpected Solution Check Status. Value '{solutionCheckStatus}' has changed from '{originalSolutionStatus}'. A failure occurred during cancellation and was not reported to the message bar. ");
+                        }
                     }
 
                     appBrowser.ThinkTime(10000);
