+++ Microsoft.PowerApps.UIAutomation.Sample/ModelDriven/VerifySolutionChecker.cs
@@ -40,70 +40,6 @@ namespace Microsoft.PowerApps.UIAutomation.Sample.ModelDriven
             _browserType = (BrowserType)Enum.Parse(typeof(BrowserType), _testContext.Properties["BrowserType"].ToString());
         }
 
-        [TestMethod]
-        public void TestVerifySolutionChecker()
-        {
-            BrowserOptions options = TestSettings.Options;
-            options.BrowserType = _browserType;
-
-            using (var appBrowser = new PowerAppBrowser(TestSettings.Options))
-            {
-                try
-                {
-                    // Login To PowerApps
-                    Console.WriteLine("Performing Login");
-
-                    for (int retryCount = 0; retryCount < Reference.Login.SignInAttempts; retryCount++)
-                    {
-                        try
-                        {
-                            appBrowser.OnlineLogin.Login(_xrmUri, _username.ToSecureString(), _password.ToSecureString());
-                            break;
-                        }
-                        catch (Exception)
-                        {
-                            appBrowser.Navigate("about:blank");
-                            if (retryCount == Reference.Login.SignInAttempts)
-                            {
-                                Console.WriteLine("Login failed after {0} attempts.", retryCount);
-                                throw;
-                            }
-                            Console.WriteLine("Login failed in #{0} attempt.", retryCount);
-                            continue;
-                        }
-                    }
-                    Console.WriteLine("Login Complete");
-
-                    //Pick the Org
-                    Console.WriteLine($"Changing PowerApps Environment to {_environmentName}");
-                    appBrowser.Navigation.ChangeEnvironment(_environmentName);
-                    appBrowser.ThinkTime(1500);
-
-                    //Click Solutions
-                    Console.WriteLine($"Click {_sideBarButton} via Sidebar");
-                    appBrowser.SideBar.Navigate(_sideBarButton);
-
-                    Console.WriteLine("Make sure each managed solution does not have the solution checker button in the command bar");
-                    appBrowser.ModelDrivenApps.VerifyManagedSolutionsUnavailable(_commandBarButton ,1000);
-
-                    appBrowser.ThinkTime(5000);
-
-                }
-                catch (Exception e)
-                {
-                    Console.WriteLine($"An error occurred during Solution Checker test run for solution {_solutionName}: {e}");
-                    string location = $@"{_resultsDirectory}\RunSolutionChecker-{_solutionName}-GenericError.bmp";
-
-                    appBrowser.TakeWindowScreenShot(location, OpenQA.Selenium.ScreenshotImageFormat.Bmp);
-                    _testContext.AddResultFile(location);
-
-                    Assert.Fail($"An error occurred during Solution Checker test run for solution {_solutionName}: {e}");
-                }
-
-                Console.WriteLine("Solution Checker Test Run Complete");
-            }
-        }
-
         private void PerformLogin(PowerAppBrowser appBrowser)
         {
             //Login To PowerApps
